Move CRX2 logic to separate file
